feat(factory): implement createCustomTestdir function
* Adds `createCustomTestdir` to facilitate custom test directory creation.
* Introduces `FactoryFn` and `TestdirFactoryOptions` types for better type safety.
* Refactors `testdir` to utilize `createCustomTestdir` for improved functionality.
* Updates utility functions to support new features and enhance file handling.
* Adds comprehensive tests for file tree creation and symbolic link handling.
